murmurhash: Fix an undefined behavior when hashing function names.
In Vulkan loader GetProcAddr handlers, for functions that are
not specified in the Vulkan registry, the loader will try loading
the name from ICDs or layers, and store the function entry into
a hashtable. The hashtable uses |murmurhash()| function to hash
Vulkan function names.

murmurhash handles data as 4-byte chunks and read one chunk at a
time; and it simply converts uint8_t pointers to uint32_t pointers
while loading data. However, the address of function name might
not be 32-bit aligned and this will cause an undefined behavior in
C99 / C11.

This change fixes the murmurhash() behavior so that it only does
pointer type casting when the memory address is 32-bit aligned,
otherwise it will load data bytewise. Note that we also handle
different endianness in |fetch_four_bytes|, though this could
usually be optimized (both gcc and clang could optimize this in
-O1).

Change-Id: I16011720198a0ee96e556855858a9909f95ec376